+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Features
+
+- [#4786](https://github.com/ignite/cli/pull/4786) Add all types to the documentation and disclaimer for multiple coin types.
+
 ## [`v29.2.1`](https://github.com/ignite/cli/releases/tag/v29.2.1)
 
 ### Changes